+++ modules/ingest-common/src/main/java/org/elasticsearch/ingest/common/FingerprintProcessor.java
@@ -1,11 +1,12 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
  */
 
-package org.elasticsearch.xpack.ingest;
+package org.elasticsearch.ingest.common;
 
 import org.elasticsearch.common.Strings;
 import org.elasticsearch.common.util.ByteUtils;
