+++ server/src/main/java/org/elasticsearch/common/network/CIDRUtils.java
@@ -1,14 +1,14 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
  */
 
-package org.elasticsearch.xpack.core.common.network;
+package org.elasticsearch.common.network;
 
 import org.elasticsearch.common.collect.Tuple;
-import org.elasticsearch.common.network.InetAddresses;
 
 import java.net.InetAddress;
 import java.util.Arrays;
