--- x-pack/plugin/ingest/build.gradle
@@ -1,30 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0 and the Server Side Public License, v 1; you may not use this file except
- * in compliance with, at your election, the Elastic License 2.0 or the Server
- * Side Public License, v 1.
- */
-
-apply plugin: 'elasticsearch.esplugin'
-apply plugin: 'elasticsearch.internal-cluster-test'
-esplugin {
-  name 'x-pack-ingest'
-  description 'Elasticsearch Expanded Pack Plugin - Ingest'
-  classname 'org.elasticsearch.xpack.ingest.IngestPlugin'
-  extendedPlugins = ['x-pack-core']
-}
-archivesBaseName = 'x-pack-ingest'
-
-dependencies {
-  compileOnly project(path: xpackModule('core'))
-  testImplementation project(path: xpackModule('core'), configuration: 'testArtifacts')
-  testImplementation project(path: ':modules:ingest-common')
-  testImplementation project(path: ':modules:lang-mustache')
-  testImplementation project(path: ':modules:geo')
-  testImplementation project(path: xpackModule('monitoring'), configuration: 'testArtifacts')
-}
-
-addQaCheckDependencies()
-
-tasks.named("testingConventions").configure { enabled = false }
